Create SOP_Scheadule.py
This Script will allow the user to reference SOP Families that are located in Linked Models and schedule them by reading their metadata and writing new data parameters such as Northing and Eastings. These parameters are hard coded in the script but can be converted to variables or inputs to use in Dynamo.